test(step2): 실패하는 인수테스트 작성
- cucumber feature 파일 station과 path로 분리
