+++ README.md
@@ -41,3 +41,22 @@ cucumber가 제공하는 기능을 적절히 활용해보세요.
 - [x] 검증 시 resopnse.jsonPath().getList("...")와 같은 부분의 각 단계를 메서드로 분리하면 재사용도 가능하고 시나리오가 더욱 명확해질 수 있다.
 - [x] 거리의 타입이 int 대신 long 인 이유는?
   - 표현할 수 있는 값의 범위가 큰 것이 무조건적인 장점일까? 저장할 데이터의 최소/최대 범위를 고려해서 자료형을 정해야 한다.
+---
+# 🚀 2단계 - 요금 조회
+
+## 요구사항 - 경로 조회 시 요금 정보 포함
+- [ ] 경로 조회 결과에 요금 정보를 포함한다.
+  - 지하철 운임은 거리비례제로 책정됩니다. (실제 경로가 아닌 최단거리 기준)
+  - 최단 거리가 계산되면 요금을 책정한다.
+```
+기본운임(10㎞ 이내) : 기본운임 1,250원
+이용 거리초과 시 추가운임 부과
+10km초과∼50km까지(5km마다 100원)
+50km초과 시 (8km마다 100원)
+```
+
+## 프로그래밍 요구사항
+- [ ] 인수 테스트 주도 개발 프로세스에 맞춰 기능을 구현한다.
+  - 인수 조건 정의 -> 인수 테스트 작성 -> 인수 테스트 충족 기능 구현
+  - 인수 조건을 주석으로 작성한다
+  - 인수 테스트 이후 기능 구현은 TDD로 진행
